Move several files into builder directory
This commit also adds some build functions to the API
documentation page.
