+++ nengo_loihi/builder/network.py
@@ -2,11 +2,11 @@ from nengo.builder.network import build_network as nengo_build_network
 from nengo.network import Network
 
 from nengo_loihi.builder import Builder
+from nengo_loihi.builder.discretize import discretize_model
+from nengo_loihi.builder.split_blocks import split_model
+from nengo_loihi.builder.validate import validate_model
 from nengo_loihi.compat import seed_network
-from nengo_loihi.discretize import discretize_model
-from nengo_loihi.split_blocks import split_model
 from nengo_loihi.splitter import Split
-from nengo_loihi.validate import validate_model
 
 
 @Builder.register(Network)
