+++ nengo_loihi/builder/split_blocks.py
