+++ nengo_loihi/builder/tests/test_split_blocks.py
