+++ nengo_loihi/hardware/interface.py
@@ -13,14 +13,14 @@ import jinja2
 from nengo.exceptions import SimulationError
 import numpy as np
 
+from nengo_loihi.builder.discretize import scale_pes_errors
 from nengo_loihi.compat import make_process_step
-from nengo_loihi.discretize import scale_pes_errors
 from nengo_loihi.hardware.allocators import OneToOne, RoundRobin
 from nengo_loihi.hardware.builder import build_board
-from nengo_loihi.nxsdk_obfuscation import d, d_func, d_get
 from nengo_loihi.hardware.nxsdk_objects import LoihiSpikeInput
 from nengo_loihi.hardware.nxsdk_shim import assert_nxsdk, nxsdk, SnipPhase, SpikeProbe
 from nengo_loihi.hardware.validate import validate_board
+from nengo_loihi.nxsdk_obfuscation import d, d_func, d_get
 from nengo_loihi.probe import LoihiProbe
 
 logger = logging.getLogger(__name__)
