+++ nengo_loihi/hardware/tests/test_interface.py
@@ -7,7 +7,7 @@ import pytest
 
 from nengo_loihi.block import Axon, LoihiBlock, Synapse
 from nengo_loihi.builder import Model
-from nengo_loihi.discretize import discretize_model
+from nengo_loihi.builder.discretize import discretize_model
 from nengo_loihi.hardware import interface as hardware_interface
 from nengo_loihi.hardware.allocators import OneToOne
 from nengo_loihi.hardware.builder import build_board
