+++ nengo_loihi/tests/test_block.py
@@ -4,7 +4,7 @@ import pytest
 
 from nengo_loihi.block import Axon, LoihiBlock, Synapse
 from nengo_loihi.builder import Model
-from nengo_loihi.discretize import discretize_model
+from nengo_loihi.builder.discretize import discretize_model
 from nengo_loihi.emulator import EmulatorInterface
 from nengo_loihi.hardware import HardwareInterface
 from nengo_loihi.inputs import SpikeInput
