+++ nengo_loihi/tests/test_conv.py
@@ -12,10 +12,10 @@ import scipy.signal
 
 import nengo_loihi
 from nengo_loihi.block import Axon, LoihiBlock, Synapse
-from nengo_loihi.builder import Model
-from nengo_loihi.compat import nengo_transforms
 from nengo_loihi import conv
-from nengo_loihi.discretize import discretize_model
+from nengo_loihi.builder import Model
+from nengo_loihi.builder.discretize import discretize_model
+from nengo_loihi.compat import nengo_transforms
 from nengo_loihi.emulator import EmulatorInterface
 from nengo_loihi.hardware import HardwareInterface
 from nengo_loihi.hardware.allocators import RoundRobin
